angular.mod(sundry)	R Documentation

Applies the modified angular transformation to a vector of counts.

Description

Applies the angular transformation to a vector of counts. A modified transformation is used that is appropriate when N < 50 and the proportion is not between 0.3 and 0.7. The transformation is given by sin^{-1}{frac{count + 0.375}{n + 0.75}} {arcsin( sqrt( (count + 0.375) / (n + 0.75) ) )}.

Usage

angular.mod(count, n)

Arguments

count 	 The numeric vector of counts.

n 	 The number(s) of observations from which the count(s) were obtained.

Value

A numeric vector.

See Also

angular

Examples

n <-25
y <- rbinom(10, n, 0.5)
y <- c(y,0,n)
p.ang.mod <- angular.mod(y, n)

[Package sundry version 0.1-0 Index]