efficiencies(sundry)	R Documentation

Computes the nonzero canonical efficiency factors between two projection matrices

Description

Computes the nonzero canonical efficiency factors between two projection matrices Q.U and Q.R.

Usage

efficiencies(Q.U, Q.R, tol = 1e-10)

Arguments

Q.U 	 A symmetric projection operator.

Q.R 	 Another symmetric projection operator.

tol 	 Values smaller than tol are considered to be zero.

Value

A vector containing the nonzero canonical efficiency factors.

See Also

eigen

Examples

    efficiencies(Q.Blocks, Q.Treats)
    eff.RC.Tr <- efficiencies(Q.RowsCols, Q.Treats)

[Package sundry version 0.4-1 Index]