ginv(sundry)	R Documentation

Computes the generalized inverse of a rectangular matrix

Description

Computes the generalized inverse of a rectangular matrix using its Singular Value Decomposition.

Usage

ginv(x, tol = .Machine$double.eps ^ 0.5)

Arguments

x 	 A real rectangular matrix.

tol 	 Singular values of x less than tol times the largest singular value are considered to be zero.

Value

A matrix containing the generalized inverse.

See Also

svd

Examples

    X <- matrix(c(1,1,1,1, 1,1,0,0, 0,0,1,1), nrow=4, ncol=3)
    XtX <- t(X) 
    ginv(XtX)

[Package sundry version 0.2-0 Index]