info.crit(sundry)	R Documentation

Computes AIC and BIC for a model.

Description

Computes Akiake and Bayesian Information Criterion for a model.

Usage

info.crit(asreml)

Arguments

asreml 	 An asreml object resulting from the fitting of a model using REML.

Value

A data frame containing the Residual degrees of freedom, AIC, BIC and log of the REML value.

See Also

reml.lrt

Examples

    info.crit(model)

[Package sundry version 0.1-0 Index]