is.zero(sundry)	R Documentation

Tests whether all elements are zero.

Description

A single-line function that tests whether all elements are zero.

Usage

is.zero(x, tol = 1e-10)

Arguments

x 	 An object whose elements are to be tested.

tol 	 The value to be regarded as close enough to zero.

Value

A logical.

Examples

y <- c(rep(0,9), 1)
is.zero(y)

[Package sundry version 0.4-0 Index]