mean.harmonic(sundry)	R Documentation

Calcuates the harmonic mean.

Description

A function to calcuate the harmonic mean of a set of nonzero numbers.

Usage

mean.harmonic(x, tol = 1e-10)

Arguments

x 	 An object from whose elements the harmonic mean is to be computed.

tol 	 The value to be regarded as close enough to zero.

Value

A numeric.

Examples

y <- c(seq(0.1,1,0.2))
mean.harmonic(y)

[Package sundry version 0.4-0 Index]