mean.op(sundry)	R Documentation

computes the mean operator matrix corresponding to a (generalized) factor.

Description

Computes the symmetric mean operator matrix corresponding to a (generalized) factor.

Usage

mean.op(factor)

Arguments

factor 	 The (generalized) factor whose means the mean operator computes from an observation-length vector.

Value

A matrix containing the symmetric matrix.

Note

A previous version included a rep argument that is no longer needed

See Also

fac.combine in package dae

Examples

    AB <- factor.combine(list(A,B))
    M.AB <- mean.op(AB)

[Package sundry version 0.4-2 Index]