plot.profiles(sundry)	R Documentation

Produces a trellis graph in which the values for an individual unit are plotted.

Description

A response variable is plotted against an explanatory varaible, such as time or distance, with plots for different combinations of the levels of the conditioning factors placed in separate panels of the trellis graph. Such plots, when produced from longitudinal data, are called profile plots. Approximate confidence intervals (y +/- 2 *se) are also plotted if the se argument is given.

Usage

plot.profiles(formula, groups.factor, se=NULL, data=NULL, 
                     type="b", black=F, title="Profile plot", xlab="", ylab="", 
                     key=T, key.title, columns=2, ...)

Arguments

formula 	 A formula that specifies the response variable (Y), the explanatory variable (X) and the conditioning factors (A, B, C, ...) for the graph. It is of the form Y ~ X | A:B:C.

groups.factor 	 The factor that indexes the different plots to be grouped in to same panel. If data is specified it must contain this factor.

se 	 A numeric that contains the standard errors for the response. If data is specified it must contain this numeric. When se is given, it is used to produce approxinate confidence intervals (y +/- 2 *se) for the response, that are also plotted.

data 	 A data frame containing the variables, factors (incuding the groups.factor) and the standard errors.

type 	 Character string telling how to plot the x and y values: "l" for lines, "b" for both points and lines.

black 	 Logical specifying that, for plots of type "b", whether the lines and symbols are to be coloured or black. Plots of type "l" are always black.

title 	 A title to be placed on the graph.

xlab 	 A label for the X-axis of each panel.

ylab 	 A label for the Y-axis of each panel.

key 	 Logical indicating whether a key (legend) is to be included.

key.title 	 The title to place on the key (legend) for the different plots in the same panel.

columns 	 The number of columns in the key or legend.

... 	 Graphical parameters that are passed to points or lines. By default, the trellis parameter list plot.symbol controls character size (cex), plotting symbol (pch), font, and color (col) for points, and the plot.line parameters control line width (lwd), line type (lty), and color (col) for lines. If plotting both points and lines (type="b"), the plot.symbol colors are used. If plotting just lines (type="l"), the plot.line line types are used.

Value

An object of class trellis, which is automatically plotted.

Warning

If the groups factor is in a data frame, the name of the data frame must be specified as part of groups.factor argument.

Note

This function uses the xyplot function to produce the graph. It is essentially an implementation of this function to produce profile plots.

See Also

xyplot, panel.xyplot.

Examples

plot.profiles(formula=Height ~ Day | Distance:Treatment, 
  groups.factor=Plot, 
  xlab="Day", ylab="Height (mm)", 
  title="Profile plot for Height", key.title="Plot")

plot.profiles(formula=R ~ Wavelength | Species, groups.factor=Rep, 
  data=Spectra.vec, title="Spectra plots", key.title="Rep", columns=6)

plot.profiles(predicted.value ~ Day | Area, groups.factor=Treatment, se=standard.error, 
   data=pred, type="l", xlab="Day", ylab="Height (mm)", ylim=c(0,25), columns=3,
   title="Plot of fitted splines (with approximate confidence intervals) for 0.5 km", 
   key.title="Treatments")

[Package sundry version 0.3-0 Index]