projdf(sundry)	R Documentation

Computes the degrees of freedom or number of nonzero canonical efficiency factors of a projection matrix

Description

Computes the numbe of nonzero canonical efficiency factors of a projection matrix.

Usage

projdf(Q, tol = 1e-10)

Arguments

Q 	 A symmetric projection operator.

tol 	 Values smaller than tol are considered to be zero.

Value

A scalar containing the number of nonzero canonical efficiency factors.

See Also

eigen

Examples

    projdf(Q.Treats)
    df.RC <- projdf(Q.RowsCols)

[Package sundry version 0.4-1 Index]