projectnonorth(sundry)	R Documentation

Computes the projection and Residual operators for two projection matrices that may be nonorthogonal

Description

A procedure to compute the projection operator for Q in P, the Residual operator for P remove Q and the nonzero canonical efficiency factors for P and Q. P and Q may be nonorthogonal.

Usage

projectnonorth(P, Q, tol = 1e-10)

Arguments

P 	 A symmetric projection operator from which Q is to be estimated.

Q 	 A symmetric projection operator whose projection in P is required.

tol 	 Values smaller than tol are considered to be zero.

Value

A list made up of the P.res, the Residual operator matrix, Q.proj, the projection operator matrix, and efficiencies, a vector containing the nonzero canonical efficiency factors.

See Also

efficiencies

Examples

    proj.PBIBD <- projectnonorth(Q.Blocks.Plots, Q.Treats)
    P.RC.Tr <- (projectnonorth(Q.RowsCols, Q.Treats))$P.res

[Package sundry version 0.3-0 Index]