randrep(sundry)	R Documentation

generates patterned uniform random numbers.

Description

generates patterned uniform random numbers..

Usage

randrep(n, each=1, times=1, sets=1)

Arguments

n 	 The number of uniform random numbers to generate.

each 	 If each is supplied, each element of random number is repeated each times. This vector is repeated again if times is also given.

times 	 the number of times to replicate the n random numbers that have been repeated each times to form a set of random numbers. There are two ways to use times. If it is a single value, the whole of the repeated random numbers is replicated that many times. If it is a vector that is the same length as the number of repeated random numbers, the result is a vector with times[1] replications of the first random number, times[2] of the second, etc. Zero is allowed in both usages; if times=0, then the length of the result is 0. It is an error if the length of times is neither 1 nor equal to the number of repeated random numbers.

sets 	 The number of sets of random numbers to generate.

Value

A vector of random numbers whose length is n * each * times * sets.

See Also

order, rep, runif

Examples

    randrep(5, each = 2, times = 3, sets = 3)

    Random.Order <- order(randrep(2, each=12), randrep(3, each=4, times=2), randrep(4, sets = 6))

[Package sundry version 0.1-0 Index]