reml.lrt(sundry)	R Documentation

Performs REML likelihood ratio test.

Description

Extracts the REML log likelhood and number of variance parameters from two asreml objects. It assumes that the second asreml object is the result of fitting a model that is a reduced version of the model for the first object. In the case where the reduced model is obtained by setting positively-constrained variance parameters in the full model to zero, the positive.zero argument should be set to TRUE so that the p-value is computed using a mixture of chi-square distributions as described in Self and Liang (1987).

Usage

reml.lrt(full.asreml, reduced.asreml, positive.zero=F)

Arguments

full.asreml 	 asreml object for the full model.

reduced.asreml 	 asreml object for the reduced model.

positive.zero 	 Indicates whether the hypothesized values for the varaince components being tested are on the boundary of the parameter space. For example, this is be true for positively-constrained variance components that, under the reduced model, are zero.

Value

A data frame containg the log of the likelihood ratio, its degrees of freedom and its p-value.

Note

The degrees of freedom for the test are computed as the difference in the number of variance parameters that are unfixed for the two models.

This procedure is only approriate when the null hypothesis is that all parameters are on the boundary of the parameter space or that all parameters are in the interior of the parameter space. Mixed cases have been discussed by Self and Liang (1987), but are not implemented here.

References

Self, S.G., and Liang, K-Y. (1987) Asymptotic Properties of Maximum Likelihood Estimators and Likelihood Ratio Tests Under Nonstandard Conditions. Journal of the American Statistical Association, 82, 605-10.

See Also

info.crit

Examples

    reml.lrt(ICV.max, ICV.red, positive.zero=T)

[Package sundry version 0.1-0 Index]